pyproject.toml
[tool.poetry]
name = "pb-portal"
version = "0.1.0"
description = ""
authors = ["vaclav-v"]
readme = "README.md"
packages = [{include = "pb_portal"}]

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"
gunicorn = "^21.2.0"
uvicorn = "^0.27.1"
python-multipart = "^0.0.9"
pydantic = "^2.6.1"
jinja2 = "^3.1.3"
aiofiles = "^23.2.1"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
greenlet = "^3.0.3"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
